+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Create collections for models provided in IServerAppConfig ahead-of-time (needed for first-time ACID transactions)
 - Pluralize all model names
+- Add `IServerAppConfig.mongooseConnectionOptions`
 
 #### v1.7.12